[ANDROSDK-1209-4] Extract cookie helper
